+++ tests/test_photo_cache.py
@@ -8,7 +8,6 @@ import pytest
 
 SAMPLE_DATE = datetime.fromisoformat("2020-11-09 18:30 -05:00")
 SAMPLE_APRICOT_DIRECTORY = PurePosixPath("/resources/photos")
-SAMPLE_LOCAL_DIRECTORY = Path("/var/tmp/photos")
 CACHE_FILE_NAME = "photo_cache.pickle"
 INITIAL_CACHE = {
     None: None,
@@ -28,7 +27,6 @@ def mock_photo_retriever(mocker):
 def photo_cache(mock_photo_retriever, tmp_path):
     """Return a photo cache."""
     return PhotoCache(
-        local_directory=SAMPLE_LOCAL_DIRECTORY,
         apricot_directory=SAMPLE_APRICOT_DIRECTORY,
         urls_to_paths=INITIAL_CACHE.copy(),
         photo_retriever=mock_photo_retriever,
@@ -76,12 +74,6 @@ def test_apricot_photo_path(photo_cache):
     )
 
 
-def test_local_photo_path(photo_cache):
-    """Test building a path to a local photo."""
-    assert photo_cache.local_photo_path("photo.jpg") == Path(
-        "/var/tmp/photos/photo.jpg"
-    )
-
 
 def test_cache_photo_none(photo_cache, paid_meetup_event, mock_photo_retriever):
     """Test caching the photo for a later Meetup event without a photo."""
@@ -89,51 +81,38 @@ def test_cache_photo_none(photo_cache, paid_meetup_event, mock_photo_retriever):
     mock_photo_retriever.get.assert_not_called()
 
 
-def test_cache_photo_later(photo_cache, later_free_meetup_event, mock_photo_retriever):
+def test_cache_photo_later(photo_cache, later_free_meetup_event, mock_photo_retriever, mocker):
     """Test caching the photo for a later Meetup event."""
     expected_apricot_path = PurePosixPath(
         "/resources/photos/AC_Mending_Monday_Test_Event_2020-11-16.jpeg"
     )
-    expected_local_path = PosixPath(
-        "/var/tmp/photos/AC_Mending_Monday_Test_Event_2020-11-16.jpeg"
-    )
+    expected_photo_name = \
+        "AC_Mending_Monday_Test_Event_2020-11-16.jpeg"
+    
+    mock_photo_retriever.get = mocker.Mock(return_value=expected_photo_name)
     assert photo_cache.cache_photo(later_free_meetup_event) == expected_apricot_path
     mock_photo_retriever.get.assert_called_once_with(
-        later_free_meetup_event.photo_url, expected_local_path
+        later_free_meetup_event.photo_url, expected_photo_name
     )
 
 
 def test_cache_photo_share(
-    photo_cache, free_meetup_event, later_free_meetup_event, mock_photo_retriever
+    photo_cache, free_meetup_event, later_free_meetup_event, mock_photo_retriever, mocker
 ):
     """Test caching the photos for Meetup events that share photos."""
     expected_apricot_path = PurePosixPath(
         "/resources/photos/AC_Mending_Monday_Test_Event_2020-11-09.jpeg"
     )
-    expected_local_path = PosixPath(
-        "/var/tmp/photos/AC_Mending_Monday_Test_Event_2020-11-09.jpeg"
-    )
+    expected_photo_name = \
+        "AC_Mending_Monday_Test_Event_2020-11-09.jpeg"
+    
+    mock_photo_retriever.get = mocker.Mock(return_value=expected_photo_name)
     assert photo_cache.cache_photo(free_meetup_event) == expected_apricot_path
     assert photo_cache.cache_photo(later_free_meetup_event) == expected_apricot_path
     mock_photo_retriever.get.assert_called_once_with(
-        later_free_meetup_event.photo_url, expected_local_path
+        later_free_meetup_event.photo_url, expected_photo_name
     )
 
-
-def test_assure_local_directory_existing(tmp_path, photo_cache):
-    """Test assuring an existing photo directory."""
-    local_photo_directory = tmp_path
-    photo_cache.local_directory = local_photo_directory
-    photo_cache.assure_local_directory()
-    assert local_photo_directory.is_dir()
-
-
-def test_assure_local_directory_new(tmp_path, photo_cache):
-    """Test creating a new photo directory."""
-    local_photo_directory = tmp_path / "photos"
-    photo_cache.local_directory = local_photo_directory
-    photo_cache.assure_local_directory()
-    assert local_photo_directory.is_dir()
 
 
 def test_persist(photo_cache, tmp_path):